requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
django-filter>=22.1
gunicorn>=20.1.0
psycopg2-binary>=2.9.3
drf-yasg>=1.21.0
